pyproject.toml
[tool.poetry]
name = "nlpiper"
version = "0.1.0"
description = "NLPiper, a lightweight package integrated with a universe of frameworks to pre-process documents."
authors = ["Tomás Osório"]
maintainers = ["Carlos Alves, Daniel Ferrari, Tomás Osório"]
license = "MIT"

readme = "README.md"

packages = [
    { include = "nlpiper" }
]

repository = "https://github.com/tomassosorio/NLPiper"
documentation = "https://github.com/tomassosorio/NLPiper/README.md"

keywords = ["NLP", "CL", "natural language processing", "computational linguistics", "parsing", "tokenizing",
    "linguistics", "language", "natural language", "text analytics", "deep-learning "]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Freely Distributable",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities"
]

[tool.poetry.dependencies]
python = "^3.8"

pydantic = "^1.8.1"

bs4 = { version = "^0.0.1", optional = true }
cyhunspell = { version = "^2.0.2", optional = true }
nltk = { version = "^3.5", optional = true }
sacremoses = { version = "^0.0.43", optional = true }
stanza = { version = "^1.3.0", optional = true }
gensim = { version = "^4.1.2", optional = true }
numpy = { version = "^1.22.2", optional = true }

[tool.poetry.extras]
bs4 = ["bs4"]
nltk = ["nltk"]
sacremoses = ["sacremoses"]
hunspell = ["cyhunspell"]
stanza = ["stanza"]
gensim = ["gensim"]
numpy = ["numpy"]

all = [
    "bs4",
    "cyhunspell",
    "nltk",
    "sacremoses",
    "stanza",
    "gensim",
    "numpy",
]

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
mypy = "^0.812"
pydocstyle = "^5.1.1"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
sphinx = "^3.5.1"
sphinx-rtd-theme = "^0.5.1"
toml = "^0.10.2"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.nlpiper]
copyright = "2021, Tomás Osório"
